fix: support float input

+++ src/renderer/src/components/Product/Create.tsx
@@ -54,7 +54,7 @@ export const Create = () => {
     let parsedValue: string | number | boolean = value;
     switch (fieldDetail?.dataType) {
       case SeriesFieldDataType.number:
-        parsedValue = parseInt(value as string);
+        parsedValue = parseFloat(value as string);
         break;
       case SeriesFieldDataType.boolean:
         parsedValue = !!value;
